Delete old searchbar and dropdown replace with the navbar components. Also removed unnecessary console log clutter
