+++ client/src/components/sub/SearchBar.jsx
@@ -1,54 +1,43 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { Link } from 'react-router';
-import Dropdown from './Dropdown.jsx';
+import {Navbar as Navigation, FormGroup, FormControl} from 'react-bootstrap'
+import {connect} from 'react-redux';
+
+import NoteReducer from '../../reducers/noteReducers';
+import RoomReducer from '../../reducers/roomReducers';
+import UserReducer from '../../reducers/userReducers';
+import EntryReducer from '../../reducers/entryReducers';
+
+import {filterKeyword} from '../../actions/entryActions';
 
 export class SearchBar extends React.Component {
-
   constructor(props) {
     super(props);
-    this.state = {
-      allClasses: []
-    };
-  }
-
-  componentWillMount() {
-    this.getClassList();
   }
 
   getUserInput (e) {
-    this.props.filterKeyword(e.target.value);
+    this.props.dispatch(filterKeyword(e.target.value));
   }
-
-  getClassList () {
-    var classList = {};
-    var result = [];
-    for (var i = 0; i < this.props.entries.length; i++) {
-      var className = this.props.entries[i].class;
-      classList[className] = className;
-    }
-    for (var key in classList) {
-      this.state.allClasses.push(classList[key]);
-    }
-  }
-
-  /* <Dropdown list={this.state.allClasses} filterClass={this.props.filterClass} /> */
 
   render() {
     return (
-      <form className="navbar-form" role="search">
-        <div className="form-group searchbar">
-          <input type="text" className="form-control" placeholder="Search!" onChange={this.getUserInput.bind(this)} />
-        </div>
-        <Dropdown list={this.state.allClasses} filterClass={this.props.filterClass} />
-      </form>
-    );
+      <li role="presentation">
+        <Navigation.Form>
+          <FormGroup>
+            <FormControl type="text" placeholder="Search" onChange={this.getUserInput.bind(this)}/>
+          </FormGroup>
+        </Navigation.Form>
+      </li>
+    )
   }
 }
 
 const mapStateToProps = (state) => {
   return {
-    ...state
+    ...state,
+    RoomReducer,
+    UserReducer,
+    NoteReducer,
+    EntryReducer
   };
 };
